requirements.txt
Flask=2.0.2
newsapi-python=0.2.6
flask-cors=3.0.10